Fix binary download issue (#277)
It was reported in #269 that y-sweet downloads the binary every time. I
was able to reproduce this locally when running the script directly,
although this may be because `require.resolve` behaves weirdly (fails
silently?) outside of an npm package.

I think the underlying issue is that we had two different code paths
that determined where to check for the binary and where to put the
binary. This collapses them into one code path, and cleans up/modernizes
the code in general.