y-sweet js client support auth endpoint (#292)
This takes a very crude approach to get the YSweetProvider supporting
reconnects. It essentially wraps the `new YSweetProvider` instantiation
logic with an async function that calls the user-provided auth endpoint
and then instantiates the YSweetProvider. It attaches an `onFailure`
handler to the provider allowing it to signal that it needs to be
recreated.

A tried several approaches, but found this to be the least invasive. It
also - believe it or not - is the most straightforward to understand. It
does, however, come with the downsides that the `createYjsProvider`
changes in two ways:

1. it's now async
2. it returns an object that is essentially a shimmed Provider that
implements most of the Provider's interface
3. the Provider itself now no longer implements the Observable
interface, instead it has an `observable` property which gets
transferred to the next Provider in case the old one gets recreated

I don't _think_ these are huge downsides as there don't seem to be a lot
of use cases where the application needs to grab the provider and do
something with it. In the case that an application _does_ need to user
the returned provider, then the calling code will need to `await`.

A more robust solution might take the approach of keeping the same
Provider but resetting the relevant bits of internal state when getting
a new client token from the auth endpoint. That approach ended up being
quite messy and error prone with old websocket event handlers and
setTimeouts hanging around. It might make sense to clean that up when we
rewrite the YSweetProvider in the coming weeks.