[y sweet/client] support provider.on() (#301)
When we [reworked the provider to support
reconnections](https://github.com/jamsocket/y-sweet/pull/292), we lost
the `on`/`off`/`once` methods from the `Observer` base class that used
to exist on `YSweetProvider`. It turns out those methods are often used
by third-party libraries. This PR adds support for those back in.