Fix debugger console.log (#327)
We want the debugger to be discoverable, but we only want to show it
once per session. We used to do this by displaying it in the `useYDoc`
hook, but that may be run more than once.

Instead, this moves the code that prints it into the `useEffect` of the
`YDocProvider`.