Use hardcoded connection string fallback rather than .env file (#348)
Seems like npm filters out `.env` files when publishing. It's probably a
good thing on the whole, but we actually want them included for
`create-y-sweet-app` because we copy them over when bootstrapping a new
project.

Rather than try to work around the issue, I just removed the `.env`
files and set up the document managers like this:

```js
const manager = new DocumentManager(
  process.env.CONNECTION_STRING || "ys://127.0.0.1:8080",
);
```

As a plus, it's a little more explicit what's happening in the code.