Migrate Python library to uv (#345)
It's fast and seems to be the direction Python projects are going. We
also now spin up a server for each test, for better isolation. This
required some changes to the piping on the server side, to allow us to
use port `0` to get the OS to assign an available port. This means we
need to compute the connection string _after_ we have bound the port, so
that the connection string contains the actual bound port and not port
`0`.

This also sets up Python tests to run in CI/CD, but only when the Python
code changes (at least for now).