Implement events for sync status (#353)
`y-websocket` has a `synced` flag, but it is a bit misleading because it
only tells whether the initial handshake was completed. Knowing whether
a document was synced _in general_ for `y-protocols` is a bit tricky,
because we would need to compare the state vector and delete set.

When using WebSockets, we can take advantage of the ordering guarantee
and synchronous message processing in Y-Sweet to implement a very simple
check of sync.

The client stores two numbers, `lastSyncSent` and `lastSyncAcked`. When
it sends an update to the server, it also increments `lastSyncSent` and
then sends a separate `messageSyncStatus` message containing that number
as the payload.

When the server receives a `messageSyncStatus`, it simply echoes it
verbatim. When the client receives the `messageSyncStatus` in return, it
updates `lastSyncAcked` to the payload of the message.

If `lastSyncAcked` = `lastSyncSent`, we know that all messages since the
last update have been processed. Otherwise, there are outstanding
changes.

This builds on #352 and will enable #306.

Demo:


https://github.com/user-attachments/assets/60d4aec2-f025-4e84-a594-28fbc84c67cb