Ensure new provider works with older server versions (#363)
The new client relies on a message type in order to support heartbeats
and sync status. If it's talking to an older server version, the server
will ignore its requests for heartbeats, and the client will assume the
connection is unhealthy and attempt to reconnect in a periodic loop.

To test this, run the examples along with an old version of the server,
e.g. `npx y-sweet@0.6.0 serve`

This fixes the library by adding a boolean field
`receivedAtLeastOneSyncResponse`. As long as this is `false`, we will
continue to send heartbeats and sync status requests, but we will never
terminate a connection over a missed heartbeat.

This is a somewhat crude approach. We should support version negotiation
eventually. But it's nice and simple, and will be easy to rip out later.