Fix debugger console logging in React hook (#369)
The new Y-Sweet provider broke the code that printed the Y-Sweet
debugger to the console, because the code that reads the active client
token runs before the client token is fetched (since fetching now
happens asynchronously). This fixes the debugger by installing an event
listener on the provider, which is triggered when the client token
becomes available.

Reported by Sorin [on
discord](https://discord.com/channels/939641163265232947/1202346195142312019/1325396209770303489).

---------

Co-authored-by: Jake Lazaroff <jake@jamsocket.com>