Fix persistence workers never terminating  (#401)
Hey!

This PR fixes an issue where memory is never reclaimed and keeps rising,
due to persistence workers never shutting down and freeing up the used
memory.

I'm running y-sweet as a long-running service, instead of via Plane or
another way to spin up a process per "room". And while experimenting I
also allowed images to be uploaded to the room state, which made the
fact that there was something up with the memory usage very visible, and
after adding a bit of logging/monitoring I could see that the number of
persistence workers never decreased.

In order to fix this, this PR adds a link between the GC Worker and the
persistence worker. When the GC detects that there's no more references
to the document it also notifies the persistence worker to shut down,
and when it does the memory is also freed.

Co-authored-by: Per-Victor Persson <pepe@netlight.com>