Allow to disable GC within documents using env variable/CLI (#422)
### References

- closes #421 

### Description

Adds an option to disable GC within documents (distinct from GC of
documents in multi-document mode) via:
- CLI by passing `--skip-gc` to either of `y-sweet serve` or `y-sweet
serve-doc`
- `Y_SWEET_SKIP_GC` environment variable

I was not sure if I should also add the environment variable to
`y-sweet-worker` for feature parity but given that it is deprecated I
went for a smaller changeset and only added it in the server.

I wonder if the configuration for toggling GC of documents themselves
(the `doc_gc` option which is not exposed by clap) would be advisable,
but I think it is already possible to configure
`checkpoint_freq_seconds` instead which can be set to a very high value
that is probably better than turning it off altogether.